+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### CURRENT
-* Added ability to open folder for download
+* Added ability to copy download URI
+* Added ability to open download folder
 * Fixed Linux shell script mistake
 * Fixed issue on Linux system with main icon being too large
 * Covered core and HTTP parts with tests